ui can generate audio
