+++ tts_audiobook_creator/epub_to_text.py
@@ -119,7 +119,7 @@ def epub_to_raw_text_book(epub_path: str | Path, max_sentence_length: int = 300)
             # Insert periods at sentence splits
             cleaned_text = insert_periods_at_splits(cleaned_text, max_length=max_sentence_length)
 
-            chapter["text"] = cleaned_text
+            chapter["body"] = cleaned_text
 
             chapters.append(chapter)
