+++ tts_audiobook_creator/main.py
@@ -1,6 +1,6 @@
 import logging
 
-from tts_audiobook_creator import Audiobook_Controller
+from tts_audiobook_creator.controller import Audiobook_Controller
 
 logging.basicConfig(level=logging.WARNING)
